+++ packages/s2-react/playground/config.ts
@@ -42,6 +42,7 @@ export const s2Options: SheetComponentOptions = {
     },
   },
   conditions: {
+    text: [],
     interval: [
       {
         field: 'number',
